--- Assets/Base/Script/Button.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 91ae20335cd09b0439820a30c581f3ad
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
