--- Assets/Base/Script/Button/ButtonGlow.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: ef82dd44f1dd3d940aa075eed1362e5a
