--- Assets/Base/Script/Button/InteractionUIButton.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 4d0b501ebdb6dd8428211370e9d6afbe
