--- Assets/Base/Script/Button/Joystick.cs
@@ -1,41 +1,0 @@
-using UnityEngine;
-using UnityEngine.EventSystems;
-
-public class Joystick : MonoBehaviour, IDragHandler, IPointerUpHandler, IPointerDownHandler
-{
-    public RectTransform background;
-    public RectTransform handle;
-
-    private Vector2 inputVector;
-
-    public float Horizontal => inputVector.x;
-    public float Vertical => inputVector.y; // ✅ 추가된 부분
-
-    public void OnDrag(PointerEventData eventData)
-    {
-        Vector2 pos;
-        if (RectTransformUtility.ScreenPointToLocalPointInRectangle(
-            background, eventData.position, eventData.pressEventCamera, out pos))
-        {
-            pos /= background.sizeDelta;
-
-            inputVector = new Vector2(pos.x * 2, pos.y * 2);
-            inputVector = Vector2.ClampMagnitude(inputVector, 1.0f);
-
-            handle.anchoredPosition = new Vector2(
-                inputVector.x * (background.sizeDelta.x / 2),
-                inputVector.y * (background.sizeDelta.y / 2));
-        }
-    }
-
-    public void OnPointerDown(PointerEventData eventData)
-    {
-        OnDrag(eventData);
-    }
-
-    public void OnPointerUp(PointerEventData eventData)
-    {
-        inputVector = Vector2.zero;
-        handle.anchoredPosition = Vector2.zero;
-    }
-}
