--- Assets/Base/Script/Button/Joystick.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 42efc31088015ec4ba97a6f75165110f
