--- Assets/Base/Script/Button/JoystickController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: b5a7fc242a98efc42b0c815fd4c08541
