--- Assets/Base/Script/Button/MainMenu.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 0ec5ed57b4947614a9b1a77f25c72812
