--- Assets/Base/Script/Button/OptionToggleUI.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 931d2166e5f295a48a1e204ff1245c4c
