--- Assets/Base/Script/Button/OptionsMenu.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 507df0bf5894e5d4a9d945f87fc8c679
