--- Assets/Base/Script/Button/StageSelector.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 383598ff00bf5d44abc8593f6894a677
