--- Assets/Base/Script/Button/SubStageUI.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 979192b941db23b44b1d0ff0ae4b655e
