+++ Assets/Base/Script/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 91ae20335cd09b0439820a30c581f3ad
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
