+++ Assets/Base/Script/UI/Button/InteractionUIButton.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 4d0b501ebdb6dd8428211370e9d6afbe
