+++ Assets/Base/Script/UI/JoystickController.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: b5a7fc242a98efc42b0c815fd4c08541
