+++ Assets/Base/Script/UI/MainMenu.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 0ec5ed57b4947614a9b1a77f25c72812
