+++ Assets/Font.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 97e75ee8faac01246a8acd7627f23e66
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
