+++ Assets/Font/HBIOS-SYS SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1069186d97a9c9e46aaa5a7ac87cebd6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
