+++ Assets/Font/Korail_Round_Gothic_Light SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 798e052be474f33458ea6e704b5ac822
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
